+++ src/routers/least_message_router.rs
@@ -5,7 +5,7 @@ use crate::actor::handler::Handler;
 use crate::prelude::{Actor, ActorMessage, ActorResult};
 use crate::routers::add_actor_message::AddActorMessage;
 use crate::routers::remove_actor_message::RemoveActorMessage;
-use log::error;
+use log::{debug, error};
 use std::error::Error;
 
 pub struct LeastMessageRouter<A>
@@ -16,6 +16,8 @@ where
     min_mailbox_size: usize,
     route_to: Vec<ActorWrapper<A>>,
     can_route: bool,
+    stop_on_system_stop: bool,
+    stop_on_empty_targets: bool,
 }
 
 /// implements [ActorFactory](../prelude/trait.ActorFactory.html) to spawn a LeastMessageRouter within an [ActorSystem](../prelude/struct.ActorSystem.html)
@@ -66,7 +68,7 @@ where
 ///     .unwrap();
 ///
 /// // create the router, fill it, and route a message
-/// let router_factory = LeastMessageRouterFactory::new(15);
+/// let router_factory = LeastMessageRouterFactory::new(15, true, true);
 /// let router = actor_system
 ///     .builder()
 ///     .spawn("router-hello-world", router_factory)
@@ -75,13 +77,21 @@ where
 /// router.send(FooBar{}).unwrap();
 /// ```
 pub struct LeastMessageRouterFactory {
-    min_mailbox_size: usize
+    /// minimum mailbox size, that needs to be exceeded to actually search for the router with the least messages
+    min_mailbox_size: usize,
+    /// defines if the actor should automatically be stopped when the system is stopped. If set to false it's up to the user to setup their own shutdown process if they want a quick and clean exit
+    stop_on_system_stop: bool,
+    /// defines if the actor should automatically be stopped if it receives a message after all targets have been automatically removed
+    /// this does not apply if the last target has been removed through a `RemoveActorMessage`
+    stop_on_empty_targets: bool,
 }
 
 impl LeastMessageRouterFactory {
-    pub fn new(min_mailbox_size: usize) -> Self {
+    pub fn new(min_mailbox_size: usize, stop_on_system_stop: bool, stop_on_empty_targets: bool,) -> Self {
         Self {
             min_mailbox_size,
+            stop_on_system_stop,
+            stop_on_empty_targets,
         }
     }
 }
@@ -94,7 +104,7 @@ where
         &mut self,
         _context: ActorContext<LeastMessageRouter<A>>,
     ) -> Result<LeastMessageRouter<A>, Box<dyn Error>> {
-        return Ok(LeastMessageRouter::new(self.min_mailbox_size));
+        return Ok(LeastMessageRouter::new(self.min_mailbox_size, self.stop_on_system_stop, self.stop_on_empty_targets));
     }
 }
 
@@ -102,17 +112,34 @@ impl<A> LeastMessageRouter<A>
 where
     A: Actor,
 {
-    pub fn new(min_mailbox_size: usize) -> Self {
+    pub fn new(min_mailbox_size: usize, stop_on_system_stop: bool, stop_on_empty_targets: bool,) -> Self {
         Self {
             next_route_index: 0,
             min_mailbox_size,
             route_to: Vec::new(),
             can_route: false,
+            stop_on_system_stop,
+            stop_on_empty_targets,
         }
     }
 }
 
-impl<A> Actor for LeastMessageRouter<A> where A: Actor {}
+impl<A> Actor for LeastMessageRouter<A> where A: Actor {
+    fn on_system_stop(&mut self, context: &ActorContext<Self>) -> Result<ActorResult, Box<dyn Error>> {
+        if self.stop_on_system_stop {
+                let result = context.actor_ref.stop();
+                if result.is_err() {
+                    error!(
+                    "Could not forward message ActorStopMessage to target {}",
+                    context.actor_ref.get_address().actor
+                );
+                return Ok(ActorResult::Stop);
+            }
+        }
+        return Ok(ActorResult::Ok);
+
+    }
+}
 
 impl<A> Handler<AddActorMessage<A>> for LeastMessageRouter<A>
 where
@@ -166,11 +193,32 @@ where
             return Ok(ActorResult::Ok);
         }
 
-        let mut target = self.route_to.get(self.next_route_index).unwrap();
+        let mut target;
+        // skip/remove stopped actors
+        loop {
+            let route_index = self.next_route_index;
+            target = self.route_to.get(self.next_route_index).unwrap();
+
+            if target.is_stopped() {
+                self.next_route_index += 1;
+                if self.next_route_index >= (self.route_to.len() - 1) {
+                    self.next_route_index = 0;
+                }
+                self.route_to.remove(route_index);
+                if self.route_to.len() == 0 && self.stop_on_empty_targets {
+                    debug!("Stopping router, because all targets have been removed");
+                    return Ok(ActorResult::Stop)
+                }
+            }
+            else {
+                break;
+            }
+        }
+
         let mut mailbox_size = target.get_mailbox_size();
         let target_len = self.route_to.len();
         for i in 0..target_len {
-            if mailbox_size >= self.min_mailbox_size {
+            if mailbox_size <= self.min_mailbox_size {
                 break;
             }
             let potential_target = self.route_to.get(i).unwrap();