+++ src/system/system_state.rs
@@ -60,13 +60,15 @@ impl SystemState {
 
     fn shutdown(&self, timeout: Duration) {
         let now = Instant::now();
+
         while self.get_actor_count() != 0 {
+
             if now.elapsed() >= timeout || self.is_force_stopped.load(Ordering::Relaxed) {
                 self.is_force_stopped.store(true, Ordering::Relaxed);
                 self.mailboxes.clear();
                 break;
             }
-            sleep(Duration::from_millis(10));
+            sleep(Duration::from_millis(100));
         }
         self.is_stopped.store(true, Ordering::Relaxed);
     }
@@ -110,24 +112,7 @@ impl SystemState {
         }
     }
 
-    pub fn remove_mailbox(&self, address: &ActorAddress) {
-        self.total_actor_count.fetch_sub(1, Ordering::Relaxed);
-        self.pool_actor_count
-            .entry(address.pool.clone())
-            .and_modify(|v| {
-                v.fetch_sub(1, Ordering::Relaxed);
-            });
-        self.mailboxes.remove(address);
-    }
-
-    pub fn add_mailbox<A>(
-        &self,
-        address: ActorAddress,
-        mailbox: Mailbox<A>,
-    ) -> Result<(), ActorError>
-    where
-        A: Handler<SerializedMessage> + 'static,
-    {
+    pub fn increase_pool_actor_count(&self, address: &ActorAddress) -> Result<(), ActorError> {
         let maximum_actor_count = self.max_actors_per_pool.get(&address.pool);
         if maximum_actor_count.is_none() {
             return Err(ActorError::ThreadPoolDoesNotExistError);
@@ -148,8 +133,34 @@ impl SystemState {
 
         current_pool_count.fetch_add(1, Ordering::Relaxed);
         self.total_actor_count.fetch_add(1, Ordering::Relaxed);
-        self.mailboxes.insert(address, Arc::new(mailbox));
+
         return Ok(());
+    }
+
+    pub fn decrease_pool_actor_count(&self, address: &ActorAddress) {
+        self.pool_actor_count
+            .entry(address.pool.clone())
+            .and_modify(|v| {
+                v.fetch_sub(1, Ordering::Relaxed);
+            });
+        self.total_actor_count.fetch_sub(1, Ordering::Relaxed);
+
+    }
+
+    pub fn remove_mailbox(&self, address: &ActorAddress) {
+        self.decrease_pool_actor_count(address);
+        self.mailboxes.remove(address);
+    }
+
+    pub fn add_mailbox<A>(
+        &self,
+        address: ActorAddress,
+        mailbox: Mailbox<A>,
+    )
+    where
+        A: Handler<SerializedMessage> + 'static,
+    {
+        self.mailboxes.insert(address, Arc::new(mailbox));
     }
 
     pub fn add_pool_actor_limit(&self, pool_name: String, max_actors: usize) {
@@ -161,11 +172,18 @@ impl SystemState {
         if maximum_actor_count.is_none() {
             return Err(ActorError::ThreadPoolDoesNotExistError);
         }
+
         let maximum_actor_count = maximum_actor_count.unwrap();
         let maximum_actor_count = *maximum_actor_count.value();
 
-        let current_pool_count = self.pool_actor_count.get(pool_name).unwrap();
-        let current_pool_count = current_pool_count.value().load(Ordering::Relaxed);
+        let current_pool_count = self.pool_actor_count.get(pool_name);
+
+        let current_pool_count = if current_pool_count.is_some() {
+            let current_pool_count = current_pool_count.unwrap();
+            current_pool_count.value().load(Ordering::Relaxed)
+        } else {
+            0 as usize
+        };
 
         if maximum_actor_count == 0 {
             let result = usize::MAX - current_pool_count;